requirements.txt
lightning==2.2.1
torchvision==0.17.2
torchmetrics==1.3.2
wandb==0.17.0
omegaconf==2.3.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
matplotlib==3.9.0